refactor: change response of /api/me endpoint and update frontend accordingly
