+++ backend/routes/usersRouter.js
@@ -5,8 +5,6 @@ import { adminRouter } from "./adminRouter.js";
 import { isAdmin } from "../auth/isAdmin.js";
 import { contributionRouter } from "./contributionRouter.js";
 
-usersRouter.get("/me", usersController.me);
-
 usersRouter.post("/logout", usersController.logout);
 
 usersRouter.use("/admin", isAdmin, adminRouter);
