+++ frontend/src/customHooks/useStatusCheck.js
@@ -12,10 +12,8 @@ function useStatusCheck(setLoading, addNotification, longWait) {
     async function checkLogin() {
       let response;
       try {
-        response = await fetch(`${BACKEND_URL}/users/me`, {
-          mode: "cors",
+        response = await fetch(`${BACKEND_URL}/api/me`, {
           method: "GET",
-          credentials: "include",
           signal: abortController.signal,
         });
